+++ history.md
@@ -1,3 +1,7 @@
+* Fire and smoke may result when terrain is damaged
+* Attacks may cause damage to terrain
+* Added AreaEnchantments
+* pos property of PlaceableThing now defaults to (0,0)
 * Fixed bug with random MissionText
 * Scenario builder automatically adds QOL reporter to cities
 * Added message log